QuickReport component added
